# Optimize `have_seen_events`