@Nira34 2022-08-19
https://datablesingles.com/tds/ae?tds_campaign=s8821lap&tdsId=s8821lap_r&s1=int&utm_source=int&utm_term=23&clickid={clickid}&subid=3X3Xvit&subid2={subid2}&affid=c3367be9

On Fri, Aug 19, 2022, 1:28 PM Eric Eastwood ***@***.***>
wrote:

> Optimize have_seen_events because when backfilling #matrixhq, 20s is just
> calling have_seen_events on the 200k state and auth events in the room.
> And in the future, hopefully a PR to avoid the have_seen_events
> altogether or do it in the background.
>
> have_seen_events (157 db.have_seen_events) takes 6.62s to process 77k
> events
> have_seen_events (246 db.have_seen_events) takes 13.19s to process 122k
> events
>
>
> <https://user-images.githubusercontent.com/558581/185563852-f9d38e8f-15f5-4ca1-8878-ef01ec95fb60.png>
>
> For a cold query, we're seeing a 3.7s -> 0.33s improvement. And the new
> version doesn't even have a cache yet and it still performs almost as good.
> Before After
> .
>
> Benchmark time (1 cold cache ): 3.7170820236206055
> Benchmark time (2, warm cache): 0.2985079288482666
> Benchmark time (3, warm cache): 0.28847789764404297
> Benchmark time (4, odds      ): 0.1537461280822754
> Benchmark time (5, odds      ): 0.14780497550964355
> Benchmark time (6, evens     ): 0.1475691795349121
> Benchmark time (7, evens     ): 0.14868617057800293
>
> .
>
> Benchmark time (1 cold cache ): 0.31160497665405273
> Benchmark time (2, warm cache): 0.3012270927429199
> Benchmark time (3, warm cache): 0.2938399314880371
> Benchmark time (4, odds      ): 0.15165090560913086
> Benchmark time (5, odds      ): 0.14775395393371582
> Benchmark time (6, evens     ): 0.14690089225769043
> Benchmark time (7, evens     ): 0.14881110191345215
>
> Part of #13356 <https://github.com/matrix-org/synapse/issues/13356>
> Dev notes
>
> SYNAPSE_POSTGRES=1 SYNAPSE_TEST_LOG_LEVEL=INFO python -m twisted.trial tests.storage.databases.main.test_events_worker.HaveSeenEventsTestCase.test_benchmark
>
> ^.* (test|sentinel) -
>
> Pull Request Checklist
>
>    - Pull request is based on the develop branch
>    - Pull request includes a changelog file
>    <https://matrix-org.github.io/synapse/latest/development/contributing_guide.html#changelog>.
>    The entry should:
>       - Be a short description of your change which makes sense to users.
>       "Fixed a bug that prevented receiving messages from other servers." instead
>       of "Moved X method from EventStore to EventWorkerStore.".
>       - Use markdown where necessary, mostly for code blocks.
>       - End with either a period (.) or an exclamation mark (!).
>       - Start with a capital letter.
>       - Feel free to credit yourself, by adding a sentence "Contributed
>       by @github_username." or "Contributed by [Your Name]." to the end of the
>       entry.
>    - Pull request includes a sign off
>    <https://matrix-org.github.io/synapse/latest/development/contributing_guide.html#sign-off>
>    - Code style
>    <https://matrix-org.github.io/synapse/latest/code_style.html> is
>    correct
>    (run the linters
>    <https://matrix-org.github.io/synapse/latest/development/contributing_guide.html#run-the-linters>
>    )
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/matrix-org/synapse/pull/13561
> Commit Summary
>
>    - 7cb07d3
>    <https://github.com/matrix-org/synapse/pull/13561/commits/7cb07d3a0338408a921120ac38d2971c312f316f>
>    Start of optimizing
>    - 2fdbca6
>    <https://github.com/matrix-org/synapse/pull/13561/commits/2fdbca62e4d69621be61963d7dece90fbc1adbcd>
>    Add benchmark
>    - fd941cd
>    <https://github.com/matrix-org/synapse/pull/13561/commits/fd941cdcbf457af4097f9f2a817cb2bf53f87c80>
>    Better benchmark
>
> File Changes
>
> (3 files <https://github.com/matrix-org/synapse/pull/13561/files>)
>
>    - *M* synapse/handlers/federation_event.py
>    <https://github.com/matrix-org/synapse/pull/13561/files#diff-a91beeaf44154c199a11f5539f2358704a1df966ff74ad8f2681b64430843f8c>
>    (6)
>    - *M* synapse/storage/databases/main/events_worker.py
>    <https://github.com/matrix-org/synapse/pull/13561/files#diff-28da6922f7ab91da0023fa1d9b4ba5d8fa3d5abe924e9160e0646c852b31f6fa>
>    (72)
>    - *M* tests/storage/databases/main/test_events_worker.py
>    <https://github.com/matrix-org/synapse/pull/13561/files#diff-056040ce4666fb61a36c8f26c38f6e6af99b4cb6c6b67f0e116ce844de10b843>
>    (68)
>
> Patch Links:
>
>    - https://github.com/matrix-org/synapse/pull/13561.patch
>    - https://github.com/matrix-org/synapse/pull/13561.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/matrix-org/synapse/pull/13561>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AYCHE6OHMSQKTZYQCSGO7JTVZ4ZSRANCNFSM567XGG4A>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>